feat: Display Environment and GitHub Payload in YML Workflow
This commit introduces an enhancement to the existing YML workflow. 📄🚀

We added a new code segment named 'display env and github payload'. This segment exports the GitHub event as an environment variable and displays all the environment variables and Github event payload when the action runs. 🌐📤

In the Introduce Member job, we updated the JSON payload in the Dispatch Repository step. 🔄 We replaced the 'event' property with 'sender' property to store the actor who triggered the event and added a new property 'issue owner' who captures the login of the issue creator. 🔑👥 This will help in providing customized handling based on the user initiating the event or issue.