fix: update badge URL in README.md
This commit addresses an incorrect URL in README.md. It updates the URL that leads to GitHub workflow's status badge for 'Intro'. As such, the URL now points to the 'Intro' page within the 'buildstuffdemo' repository rather than the 'welcome' page, ensuring that the badge accurately reflects the status of the workflow. 🛠️🔗