feat: Enhance issue body handling and update user instructions
This commit refactors the way the issue body is handled in the Github action workflow. The erstwhile method of setting issue body as an output and replacing newlines has been optimized into a single step, thereby shortening and simplifying the code. 😌📄✨

The Github action now compresses the issue body by replacing newline characters with spaces using the `tr` command instead of `sed`. This change, summarized as "Compress Body", eliminates two steps and is more efficient. 🚀

Additionally, a user instruction has been added to notify users that the invitation to join the organization might take a minute or two to show up and suggest refreshing the page if it doesn't appear immediately. These instructions have been included in both the Github action and README. 🙋‍♀️📝