feat: Update README with actionable URLs for ease of navigation
This change enhances the developer experience on README by replacing plain text instructions with hyperlinks that directly lead to the relevant pages. Now, you just need to click on the link to accept the invitation to the buildstuffdemo org or create a new repository within this org. It's now quicker, simpler and less error-prone! 🚀💡